route-generator-api/requirements.txt
fastapi
uvicorn[standard]
httpx
anthropic
python-dotenv
pydantic
